use parseVideo in diff files
