+++ lib/v2/bloomberg/react-renderer.js
@@ -1,6 +1,34 @@
 const art = require('art-template');
 const path = require('path');
-const { processVideo } = require('./utils');
+const got = require('@/utils/got');
+const { headers } = require('./utils');
+
+const processVideo = async (bmmrId, summary) => {
+    const api = `https://www.bloomberg.com/multimedia/api/embed?id=${bmmrId}`;
+    const res = await got(api, { headers });
+
+    // Blocked by PX3, return the default
+    const redirectUrls = res.redirectUrls.map(String);
+    if (redirectUrls.some((r) => new URL(r).pathname === '/tosv2.html')) {
+        return {
+            stream: '',
+            mp4: '',
+            coverUrl: '',
+            caption: summary,
+        };
+    }
+
+    if (res.data) {
+        const video_json = res.data;
+        return {
+            stream: video_json.streams ? video_json.streams[0]?.url : '',
+            mp4: video_json.downloadURLs ? video_json.downloadURLs['600'] : '',
+            coverUrl: video_json.thumbnail?.baseUrl ?? '',
+            caption: video_json.description || video_json.title || summary,
+        };
+    }
+    return {};
+};
 
 const nodeRenderers = {
     paragraph: async (node, nextNode) => `<p>${await nextNode(node.content)}</p>`,
@@ -188,4 +216,5 @@ const documentToHtmlString = async (document) => {
 
 module.exports = {
     documentToHtmlString,
+    processVideo,
 };
