+++ lib/v2/bloomberg/utils.js
@@ -6,7 +6,7 @@ const { parseDate } = require('@/utils/parse-date');
 const got = require('@/utils/got');
 const { art } = require('@/utils/render');
 
-const { documentToHtmlString } = require('./react-renderer');
+const { documentToHtmlString, processVideo } = require('./react-renderer');
 
 const rootUrl = 'https://www.bloomberg.com/feeds';
 const sel = 'script[data-component-props="ArticleBody"], script[data-component-props="FeatureBody"]';
@@ -346,33 +346,6 @@ const processBody = async (body_html, story_json) => {
     }
 
     return $.html();
-};
-
-const processVideo = async (bmmrId, summary) => {
-    const api = `https://www.bloomberg.com/multimedia/api/embed?id=${bmmrId}`;
-    const res = await got(api, { headers });
-
-    // Blocked by PX3, return the default
-    const redirectUrls = res.redirectUrls.map(String);
-    if (redirectUrls.some((r) => new URL(r).pathname === '/tosv2.html')) {
-        return {
-            stream: '',
-            mp4: '',
-            coverUrl: '',
-            caption: summary,
-        };
-    }
-
-    if (res.data) {
-        const video_json = res.data;
-        return {
-            stream: video_json.streams ? video_json.streams[0]?.url : '',
-            mp4: video_json.downloadURLs ? video_json.downloadURLs['600'] : '',
-            coverUrl: video_json.thumbnail?.baseUrl ?? '',
-            caption: video_json.description || video_json.title || summary,
-        };
-    }
-    return {};
 };
 
 const asyncPoolAll = async (...args) => {
